requirements.txt
aiohttp==1.0.5
astroid==1.6.5
async-timeout==3.0.0
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
discord==0.0.2
discord.py==0.16.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.3.1
pychalk==2.0.0
pycparser==2.18
pylint==1.9.2
PyNaCl==1.0.1
six==1.11.0
websockets==3.4
wrapt==1.10.11
youtube-dl==2018.6.19
